+++ CHANGELOG.md
@@ -21,6 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release on PyPI.
 
-[Unreleased]: https://github.com/fedejaure/mdns-beacon/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/fedejaure/mdns-beacon/compare/v0.2.0...develop
 [0.2.0]: https://github.com/fedejaure/mdns-beacon/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/fedejaure/mdns-beacon/compare/releases/tag/v0.1.0